infra/lambda-pdf-ocr/requirements.txt
boto3>=1.35.0
pillow>=11.0.0
pytesseract>=0.3.10
pymupdf>=1.24.0
